+++ src/Symfony/Component/Serializer/Context/Normalizer/ObjectNormalizerContextBuilder.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Serializer\Context\Normalizer;
 
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+
 /**
  * A helper providing autocompletion for available ObjectNormalizer options.
  *
@@ -18,4 +20,12 @@ namespace Symfony\Component\Serializer\Context\Normalizer;
  */
 final class ObjectNormalizerContextBuilder extends AbstractObjectNormalizerContextBuilder
 {
+    /**
+     * Configures whether to throw when the property accessor refuses to write
+     * an attribute the metadata reports as writable, instead of ignoring it.
+     */
+    public function withThrowOnInaccessibleAttributes(?bool $throwOnInaccessibleAttributes): static
+    {
+        return $this->with(ObjectNormalizer::THROW_ON_INACCESSIBLE_ATTRIBUTES, $throwOnInaccessibleAttributes);
+    }
 }